consumer/requirements.txt
aiokafka==0.7.0
aiopg==1.1.0
pytest-asyncio==0.14.0
flake8==3.8.4
